pyproject.toml
[tool.poetry]
name = "exam-project"
version = "0.1.0"
description = ""
authors = ["Lars Riel <larsriel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
python-dotenv = "^1.0.1"
python-dateutil = "^2.8.2"
django-rq = "^2.10.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
